Move BlockStatus + store to libsawtooth
Moves the `BlockStatus` and `BlockStatusStore` structs from the
validator to libsawtooth. Also removes the Python FFI code for
`BlockStatus`, since it is not used.
