+++ libsawtooth/src/journal/block_wrapper.rs
@@ -1,0 +1,31 @@
+/*
+ * Copyright 2018 Intel Corporation
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ * ------------------------------------------------------------------------------
+ */
+
+#[derive(Debug, Clone, PartialEq, Eq)]
+pub enum BlockStatus {
+    Unknown = 0,
+    Invalid = 1,
+    Valid = 2,
+    Missing = 3,
+    InValidation = 5,
+}
+
+impl Default for BlockStatus {
+    fn default() -> Self {
+        BlockStatus::Unknown
+    }
+}
