+++ libsawtooth/src/journal/mod.rs
@@ -14,6 +14,8 @@
 
 pub const NULL_BLOCK_IDENTIFIER: &str = "0000000000000000";
 
+pub mod block_validator;
+pub mod block_wrapper;
 pub mod chain;
 pub mod chain_id_manager;
 pub mod fork_cache;
